+++ cTools/libs/binPrinter/machoPrinter/macho64PrinterRelocations.c
@@ -1,3 +1,27 @@
+/***
+ * MIT License
+ *
+ * Copyright (c) 2026 Konychev Valerii
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 #include "mem.h"
 #include "macho64Printer.h"
 
@@ -110,7 +134,7 @@ static
 void macho64PrintRelocInfo(const Macho64File *mf, const MachoRelocInfo *relInfo)
 {
     printf("%08"PRIx32" %-5s %-7s %-6s %-9s %-15s",
-        relInfo->r_address,
+        (uint32_t)relInfo->r_address,
         relInfo->r_pcrel ? "True" : "False",
         macho64GetRelLengthStr(relInfo->r_length),
         relInfo->r_extern? "True" : "False",
